+++ NEWS.md
@@ -22,7 +22,7 @@
 * Slope selector table for half life adjustments uses time to choose the point of interest. Also it is aesthetics have been polished (#956)
 * Partial interval parameters section now allows other calculations than `AUCINT`, such as `RCAMINT`, `AUCINTD` or `CAVGINT` among others (#524)
 * Slope selector plots count with grouping options (#333)
-* Add x/y axis limits for the exploration plots (#817)
+* Add x/y axis limits for the exploration plots (#817) and facet titles including subject count (#894)
 
 ## Bugs fixed
 * ZIP folder with results will now include the exploration tab outputs: individual plots, mean plots (#794)